Preliminary windows support
